B-A-P/pyproject.toml
[tool.poetry]
name = "ai-analytics-platform"
version = "1.0.0"
description = "Enterprise-grade, AI-powered business-analytics SaaS"
authors = ["Your Team <team@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.37"}
asyncpg = "^0.30.0"

redis = {extras = ["hiredis"], version = "^5.0.0"}
pydantic = "^2.7.2"
pydantic-settings = "^2.3.0"
httpx = "^0.28.1"
google-genai = "^1.33.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
structlog = "^24.1.0"
pandas = "^2.2.2"
polars = "^0.20.15"
openpyxl = "^3.1.5"
python-dotenv = "^1.0.1"
prometheus-client = "^0.20.0"
celery = {extras = ["redis"], version = "^5.4.0"}
apscheduler = "^3.10.4"
numpy = "^1.26.4"
python-multipart = "^0.0.9"
orjson = "^3.10.0"
alembic = "^1.18.4"
psycopg2-binary = "^2.9.11"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.28.1"
pytest-cov = "^5.0.0"
pytest-benchmark = "^4.0.0"
black = "^24.4.2"
ruff = "^0.4.8"
mypy = "^1.10.0"
pre-commit = "^3.7.0"
faker = "^25.2.0"
locust = "^2.28.0"
aiosqlite = "^0.22.1"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
explicit_package_bases = true

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
